% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbcti_package_function.R
\docType{class}
\name{cell_trajectory-class}
\alias{cell_trajectory-class}
\alias{cell_trajectory}
\title{Define cell_trajectory class}
\value{
define a cell trajectory class
}
\description{
Define cell_trajectory class
}
\section{Slots}{

\describe{
\item{\code{raw_data}}{data.frame.}

\item{\code{normalized_data}}{data.frame.}

\item{\code{cell_annotation}}{data.frame.}

\item{\code{nmf_result}}{list.}

\item{\code{selected_feature}}{list.}

\item{\code{specified_gene}}{character.}

\item{\code{tsne_data}}{list.}

\item{\code{contour_plot}}{list.}

\item{\code{distribution_estimation}}{list.}

\item{\code{point_possibility}}{list.}

\item{\code{connect_cluster}}{list.}

\item{\code{trajectory}}{list.}

\item{\code{pseudotime}}{list.}

\item{\code{trajectory_plot}}{list.}
}}

