% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbcti_package_function.R
\name{connect_cluster}
\alias{connect_cluster}
\title{Connect cluster}
\usage{
connect_cluster(object, sum_cri = 0.8, diff_cri = 0.5, vague_cri = 0.01)
}
\arguments{
\item{object}{a cell_trajectory object}

\item{sum_cri}{criteria set for sum of a data point belonging to two distribution}

\item{diff_cri}{criteria set for difference between a data point belonging to two distribution}

\item{vague_cri}{criteria set for the ratio of vague point in two distribution}
}
\value{
values in the connect_cluster slot of cell_trajectory object
}
\description{
Connect cluster
}
\examples{
connect_cluster(sample_data)
}
