% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbcti_package_function.R
\name{contour_plot}
\alias{contour_plot}
\title{Title}
\usage{
contour_plot(object, n = 1000)
}
\arguments{
\item{object}{a cell_trajectory object}

\item{n}{number of grid points in each direction}
}
\value{
values in the contour_plot slot of cell_trajectory object
}
\description{
Title
}
\examples{
contour_plot(sample_data)
}
