% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbcti_package_function.R
\name{create_object}
\alias{create_object}
\title{create cell_trajectory object}
\usage{
create_object(data, normalized = FALSE)
}
\arguments{
\item{data}{single-cell RNA-seq data frame with genes in row and cells in column}

\item{normalized}{indication of if data being normalized}
}
\value{
a cell_trajectory object
}
\description{
create cell_trajectory object
}
\examples{
create_object(sample_data_df,  normalized = FALSE)
}
