% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbcti_package_function.R
\name{data_nmf}
\alias{data_nmf}
\title{Select features via NMF step 1}
\usage{
data_nmf(object, use_normalized_data = TRUE, rank = 10, thread = 1)
}
\arguments{
\item{object}{a cell_trajectory object}

\item{use_normalized_data}{if or not use normalized data as input}

\item{rank}{rank being remained for the decomposed matrix}

\item{thread}{thread number used for parallel computation}
}
\value{
values in the nmf_result slot of cell_trajectory object
}
\description{
Select features via NMF step 1
}
\examples{
data_nmf(sample_data, use_normalized_data = TRUE, thread = 1)
}
