% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbcti_package_function.R
\name{filter_data}
\alias{filter_data}
\title{filter data but not perform normalization}
\usage{
filter_data(object, gene_cri = 0, cell_cri = 0, use_normalized_data = FALSE)
}
\arguments{
\item{object}{a cell_trajectory object}

\item{gene_cri}{gene_cri criteria for the sum of counts of any gene in all cells to be greater than}

\item{cell_cri}{criteria for cells containing genes counts greater than}

\item{use_normalized_data}{if or not use normalized data as input}
}
\value{
values in the normalized_data slot of cell_trajectory object
}
\description{
filter data but not perform normalization
}
\examples{
filter_data(sample_data, gene_cri = 0, cell_cri = 0, use_normalized_data = TRUE)
}
