% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbcti_package_function.R
\name{fr}
\alias{fr}
\title{Title}
\usage{
fr(x, y = NA, scalefac = 1, method = "covMcd", noise, gateName = "fitNorm")
}
\arguments{
\item{x}{refer to prada}

\item{y}{refer to prada}

\item{scalefac}{refer to prada}

\item{method}{refer to prada}

\item{noise}{refer to prada}

\item{gateName}{refer to prada}
}
\value{
fitted distribution parameters
}
\description{
Title
}
