% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbcti_package_function.R
\name{infer_trajectory}
\alias{infer_trajectory}
\title{Title}
\usage{
infer_trajectory(object, iter_n = 50)
}
\arguments{
\item{object}{a cell_trajectory object}

\item{iter_n}{iteration number}
}
\value{
values in the trajectory slot of cell_trajectory object
}
\description{
Title
}
\examples{
infer_trajectory(sample_data, iter_n = 50)
}
