% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbcti_package_function.R
\name{normalize}
\alias{normalize}
\title{Normalize the raw counts data}
\usage{
normalize(object, gene_cri = 0, cell_cri = 0, scale_factor = 10000)
}
\arguments{
\item{object}{a cell_trajectory object}

\item{gene_cri}{criteria for the sum of counts of any gene in all cells to be greater than}

\item{cell_cri}{criteria for cells containing genes counts greater than}

\item{scale_factor}{a scale factor for normalization}
}
\value{
values in the normalized_data slot of cell_trajectory object
}
\description{
Normalize the raw counts data
}
\examples{
normalize(sample_data, gene_cri = 0, cell_cri = 0, scale_factor = 10000)
}
