% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbcti_package_function.R
\name{plot_trajectory}
\alias{plot_trajectory}
\title{construct trajectory plot for cell_trajectory object}
\usage{
plot_trajectory(object, plot_title = "", width = 0.2, height = 0.2)
}
\arguments{
\item{object}{a cell_trajectory object}

\item{plot_title}{title of the plot}

\item{width}{jitter width}

\item{height}{jitter height}
}
\value{
values in the trajectory_plot slot of cell_trajectory object
}
\description{
construct trajectory plot for cell_trajectory object
}
\examples{
plot_trajectory(sample_data)
}
