% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbcti_package_function.R
\name{point_possibility}
\alias{point_possibility}
\title{Calculate point possibility}
\usage{
point_possibility(object, r)
}
\arguments{
\item{object}{a cell_trajectory object}

\item{r}{distance from the data point for the counting area}
}
\value{
values in the point_possibility slot of cell_trajectory object
}
\description{
Calculate point possibility
}
\examples{
point_possibility(sample_data , r = 2)
}
