% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_data.R
\docType{data}
\name{sample_data}
\alias{sample_data}
\title{Sample Data for demonstration}
\format{
A cell trajectory object
}
\usage{
sample_data
}
\description{
A cell trajectory object for demonstration
}
\concept{data}
\keyword{datasets}
