% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_data_df.R
\docType{data}
\name{sample_data_df}
\alias{sample_data_df}
\title{Sample Dataframe for demonstration}
\format{
A dataframe
}
\usage{
sample_data_df
}
\description{
A dataframe with 30 rows and 50 columns filled with random generated numbers between 0 and 10
for demonstration
}
\concept{data}
\keyword{datasets}
