% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbcti_package_function.R
\name{select_cor_feature}
\alias{select_cor_feature}
\title{Select features based on correlation}
\usage{
select_cor_feature(object, feature, use_normalized_data = FALSE, k)
}
\arguments{
\item{object}{a cell_trajectory object}

\item{feature}{the specifeid features}

\item{use_normalized_data}{if or not use normalized data as input}

\item{k}{number of highest correlation genes for each target}
}
\value{
values in the selected_feature slot of cell_trajectory object
}
\description{
Select features based on correlation
}
\examples{
select_cor_feature(sample_data, c('1','2'), use_normalized_data = FALSE, k = 5)
}
