% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbcti_package_function.R
\name{select_var_feature}
\alias{select_var_feature}
\title{Select features based on variation}
\usage{
select_var_feature(object, use_normalized_data = FALSE, n = 2000)
}
\arguments{
\item{object}{a cell_trajectory object}

\item{use_normalized_data}{if or not use normalized data as input}

\item{n}{number of features with the highest variation}
}
\value{
values in the selected_feature slot of cell_trajectory object
}
\description{
Select features based on variation
}
\examples{
select_var_feature(sample_data, use_normalized_data = FALSE, n = 10)
}
