% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbcti_package_function.R
\name{tsneplot}
\alias{tsneplot}
\title{Tsne plot for cell_trajectory object}
\usage{
tsneplot(
  object,
  perplexity = 30,
  max_iter = 500,
  specified_gene = FALSE,
  use_normalized_data = TRUE,
  pca = TRUE,
  check_duplicates = FALSE,
  title = "",
  file = ""
)
}
\arguments{
\item{object}{a cell_trajectory object}

\item{perplexity}{perplexity value for tsne}

\item{max_iter}{max iteration number}

\item{specified_gene}{feature names if specified}

\item{use_normalized_data}{if or not use normalized data as input}

\item{pca}{if pca will be performed}

\item{check_duplicates}{if or not duplicates are checked at tsne step}

\item{title}{title of plot, default as none}

\item{file}{file the plot to be written to, defalut as empity}
}
\value{
values in the tsne_data slot of cell_trajectory object
}
\description{
Tsne plot for cell_trajectory object
}
\examples{
tsneplot(sample_data, use_normalized_data = TRUE, perplexity = 5)
}
